// types.ts
export enum UserRole {
  ADMIN = 'ADMIN',
  USER = 'USER',
  GUEST = 'GUEST'
}

export interface User {
  id: string;
  name: string;
  email: string;
  role: UserRole;
  department: string; // Attribute for ABE logic
}

export enum FileStatus {
  LOCKED = 'LOCKED',
  PENDING = 'PENDING',
  APPROVED = 'APPROVED', // Ready for key entry
  UNLOCKED = 'UNLOCKED'  // Decrypted locally
}

export interface SecureFile {
  id: string;
  name: string;
  size: string;
  uploadDate: string;
  owner: string;
  sensitivityLevel: 'Low' | 'Medium' | 'High' | 'Top Secret';
  description: string;
  encryptedKey?: string; // The "secret key" generated by admin
}

export interface AccessRequest {
  id: string;
  fileId: string;
  userId: string;
  userName: string;
  status: 'PENDING' | 'APPROVED' | 'REJECTED';
  requestDate: string;
  generatedKey?: string; // The key given to user upon approval
}

export interface ChatMessage {
  role: 'user' | 'model';
  text: string;
  timestamp: Date;
}